feat: Add support for nullable serviceIds in updateUserBody schema
This commit updates the updateUserBody schema in the userSchema.ts file to include support for nullable serviceIds. The serviceIds property is now defined as an array that can contain null values. This change allows for more flexibility in updating user data by allowing the serviceIds field to be optional or empty.

Note: The commit message has been generated based on the provided code changes and recent commits.