feat: Add photo property to userSchema
This commit adds the "photo" property to the userSchema file. The "photo" property is currently empty and needs to be defined with the appropriate type. This change allows for storing and retrieving user photos in the application.

Note: The commit message has been generated based on the provided code changes and recent commits.